Move client tests to a separate package
